% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesynergy-package.R, R/bayesynergy.R
\docType{package}
\name{bayesynergy}
\alias{bayesynergy-package}
\alias{bayesynergy}
\title{bayesynergy}
\usage{
bayesynergy(
  y,
  x,
  type = 3,
  drug_names = NULL,
  experiment_ID = NULL,
  units = NULL,
  lower_asymptotes = T,
  heteroscedastic = T,
  bayes_factor = F,
  robust = F,
  rho = 0.9,
  pcprior = F,
  pcprior_hypers = c(1, 0.1, 1, 0.2),
  lambda = 0.005,
  nu = 1.5,
  method = "sampling",
  control = list(),
  ...
)
}
\arguments{
\item{y}{vector or matrix of viability measures. Replicates can be given in long or wide format.}

\item{x}{two-column matrix of drug concentrations.}

\item{type}{integer; the type of model used. Must be one of the following: 1 (Splines), 2 (GP with squared exponential kernel), 3 (GP with Matérn kernel) or 4 (GP with rational quadratic kernel).}

\item{drug_names}{vector of size 2; names of the drugs utilized for the experiment.}

\item{experiment_ID}{character; identifier of experiment, typically name of cell Line.}

\item{units}{vector of size 2; concentration units for the drugs, e.g. c("\eqn{\mu}M","\eqn{\mu}M")}

\item{lower_asymptotes}{logical; if TRUE the model will estimate the lower asymptotes of monotherapy curves.}

\item{heteroscedastic}{logical; if TRUE, the model will assume heteroscedastic measurement error.}

\item{bayes_factor}{logical; if TRUE, the Bayes factor is computed between the full model, and a model containing only the non-interaction surface.}

\item{robust}{logical: if TRUE, the model assumes a log-Pareto-tailed Normal (LPTN) distribution}

\item{rho}{numeric: the rho hyperparameter for the robust likelihood distribution}

\item{pcprior}{logical: if TRUE, the model uses a penalized complexity (PC) prior on the kernel hyperparameters of the Matérn kernel (type=3)}

\item{pcprior_hypers}{vector of size 4 giving hyperparameters for the PC prior on Matérn hyperparameters}

\item{lambda}{numeric; the parameter controls the residual noise observed in the heteroscedastic model when f = 0.}

\item{nu}{numeric; the nu parameter for the Matérn kernel. Must be one of (0.5, 1.5, 2.5)}

\item{method}{The method of estimation. Must be one of {`sampling`,`vb`} corresponding to full sampling, or variational Bayes.}

\item{control}{list; passed on to the stan sampler, e.g. for setting adapt_delta.}

\item{...}{Arguments passed to \code{\link[rstan:sampling]{rstan::sampling}} or \code{\link[rstan:vb]{rstan::vb}} (e.g. iter, chains).}
}
\value{
An object of S3 class "\code{bayesynergy}", which is a list with the following entries
\tabular{ll}{
stanfit \tab An object of class \code{\link[rstan:stanmodel-class]{stanmodel}}, returned from the sampler. \cr
posterior_mean \tab A list containing the posterior means of model parameters. \cr
data \tab A list containing the original data used to fit the model. \cr
model \tab A list containing model specification for the model fit. \cr
returnCode \tab numeric; non-zero values indicate model was fit with errors or warnings \cr
LPML \tab numeric; The log pseudo-marginal likelihood of the fitted model. \cr
divergent \tab numeric; The number of divergent transitions from the Hamiltonian sampler. \cr
messages \tab character; any warnings are collected here for inspection. \cr
bayesfactor \tab numeric; The Bayes factor comparing the full model to a simpler model containing only the non-interaction assumption, p0. A large number indicating that an interaction effect is present. See the package vignette for more information.
}
}
\description{
The package implements a Gaussian process (GP) - based regression model to estimate the response surface of the viability responses obtained when two drugs are combined. The input data consists of the measured viability responses and the doses used in the combination experiment. Furthermore, the user can choose between different kernels for the definition of the covariance structure.

The function \code{bayesynergy} is the main function of the bayesynergy package. It will fit a Bayesian semi-parametric model for in-vitro drug combination experiments to estimate synergistic and antagonistic effects.
}
\examples{
\dontrun{
library(bayesynergy)
data("mathews_DLBCL")
y_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[1]]
x_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[2]]
fit <- bayesynergy(y_mat,x_mat)
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://ocbe-uio.github.io/bayesynergy/}
}

}
\author{
\strong{Maintainer}: Leiv Rønneberg \email{leiv.ronneberg@mrc-bsu.cam.ac.uk}

Authors:
\itemize{
  \item Andrea Cremaschi
  \item Robert Hanes
  \item Chi Zhang \email{chi.zhang@medisin.uio.no}
  \item Manuela Zucknick \email{manuela.zucknick@medisin.uio.no}
}

}
