% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesynergy3.R
\name{bayesynergy3}
\alias{bayesynergy3}
\title{Bayesian semi-parametric modelling for in-vitro drug combination experiments}
\usage{
bayesynergy3(
  y,
  x,
<<<<<<< HEAD
  type = 3,
=======
>>>>>>> 7a52fb0 (updated documentation)
  drug_names = NULL,
  experiment_ID = NULL,
  units = NULL,
  lower_asymptotes = T,
  lambda = 0.005,
  method = "sampling",
  control = list(),
  ...
)
}
\arguments{
\item{y}{vector or matrix of viability measures. Replicates can be given in long or wide format.}

\item{x}{three-column matrix of drug concentrations.}

<<<<<<< HEAD
\item{type}{integer; the type of model used. Must be one of the following: 1 (Splines), 2 (GP with squared exponential kernel), 3 (GP with Matérn kernel) or 4 (GP with rational quadratic kernel).}

=======
>>>>>>> 7a52fb0 (updated documentation)
\item{drug_names}{vector of size 2; names of the drugs utilized for the experiment.}

\item{experiment_ID}{character; identifier of experiment, typically name of cell Line.}

\item{units}{vector of size 2; concentration units for the drugs, e.g. c("\eqn{\mu}M","\eqn{\mu}M")}

\item{lower_asymptotes}{logical; if TRUE the model will estimate the lower asymptotes of monotherapy curves.}

\item{lambda}{numeric; the parameter controls the residual noise observed in the heteroscedastic model when f = 0.}

<<<<<<< HEAD
=======
\item{method}{The method of estimation. Must be one of {`sampling`,`vb`} corresponding to full sampling, or variational Bayes.}

>>>>>>> 7a52fb0 (updated documentation)
\item{control}{list; passed on to the stan sampler, e.g. for setting adapt_delta.}

\item{...}{Arguments passed to \code{\link[rstan:sampling]{rstan::sampling}} or \code{\link[rstan:vb]{rstan::vb}} (e.g. iter, chains).}
}
\value{
An object of S3 class "\code{bayesynergy3}", which is a list with the following entries
\tabular{ll}{
stanfit \tab An object of class \code{\link[rstan:stanmodel-class]{stanmodel}}, returned from the sampler. \cr
posterior_mean \tab A list containing the posterior means of model parameters. \cr
data \tab A list containing the original data used to fit the model. \cr
model \tab A list containing model specification for the model fit. \cr
}
}
\description{
The function \code{bayesynergy3} is the three drug extension of \code{\link[bayesynergy:bayesynergy]{bayesynergy}}. It will fit a Bayesian semi-parametric model for in-vitro drug combination experiments to estimate synergistic and antagonistic effects for three drug combinations.
}
