% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.BayeSyneRgy.R
\name{plot.BayeSyneRgy}
\alias{plot.BayeSyneRgy}
\title{Plot function for a BayeSyneRgy object}
\usage{
\method{plot}{BayeSyneRgy}(x, add_contour = TRUE, contour_levels = 0.5,
  plot_type = "2D", save_plot = FALSE, ...)
}
\arguments{
\item{x}{BayeSyn object, the result of \code{\link{BayeSyneRgy}}.}

\item{add_contour}{logical; if TRUE, draws contour lines at specified levels}

\item{contour_levels}{the contour levels}

\item{plot_type}{which plots to display. Either "2D" for two-dimensional contour plots, "3D" for interactive three-dimensional plots of the surfaces, or "both" for both these options.}

\item{save_plot}{logical; if TRUE, plots are saved locally}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function for plotting synergy surfaces and summary statistics of a \code{BayeSyneRgy} object.
}
\details{
This function extends \code{plot} to draw response and interaction surfaces of the fitted model. Both three-dimensional interactive plots, and two-dimensional contour plots can be displayed.
}
