% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.synergyscreen.R
\name{plot.synergyscreen}
\alias{plot.synergyscreen}
\title{Plot function for a synergyscreen object}
\usage{
\method{plot}{synergyscreen}(
  x,
  groupbyExperimentID = T,
  save_plots = FALSE,
  path = NULL,
  plotdevice = "pdf",
  ...
)
}
\arguments{
\item{x}{An object of class \code{synergyscreen}, the result of \code{\link{synergyscreen}}.}

\item{groupbyExperimentID}{logical; if TRUE, individual plots are produced per experiment ID.}

\item{save_plots}{logical; if TRUE plots are saved locally.}

\item{path}{string; path for saving plots, if NULL defaults to work directory.}

\item{plotdevice}{string; device for saving plots locally, must be 'pdf' or 'png'}

\item{...}{further arguments passed on to device for plotting, useful for setting width and height of saved plots.}
}
\description{
A function for plotting summary statistics of a \code{synergyscreen} object.
}
\details{
This function extends \code{plot} to summarize results from large screens.
}
\examples{
\dontrun{
library(bayesynergy)
data("mathews_DLBCL")
experiment1 = list(y = mathews_DLBCL[[1]][[1]],
x = mathews_DLBCL[[1]][[2]],
drug_names = c("ispinesib","ibrutinib"))
experiment2 = list(y = mathews_DLBCL[[2]][[1]],
x = mathews_DLBCL[[2]][[2]],
drug_names = c("canertinib","ibrutinib"))
experiments = list(experiment1,experiment2)
fit <- synergyscreen(experiments)
plot(fit)
}

}
