% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.BayeSyneRgy.R
\name{predict.BayeSyneRgy}
\alias{predict.BayeSyneRgy}
\title{Predict function for BayeSyneRgy object}
\usage{
\method{predict}{BayeSyneRgy}(object, x_mat_new, log10_conc = FALSE,
  cred_level = 0.95, plot_pred = FALSE, ...)
}
\arguments{
\item{object}{The fitted model from \code{\link{BayeSyneRgy}}.}

\item{x_mat_new}{Matrix of concentrations to predict.}

\item{log10_conc}{logical; if true, concentrations must be given on the log10 scale, and monotherapies must have value -Inf for the drug with concentration zero.}

\item{cred_level}{The credible level for equal-tailed credible intervals.}

\item{plot_pred}{logical; if true, posterior predictive densities are plotted.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function for predicting response and interaction at an unobserved location using interpolation from the Bayesian semi-parametric model.
}
\details{
This function extends \code{predict} to the \code{BayeSyneRgy} class. The function provides point estimates and uncertainty quantifiaction for new points on the surface, as specified in the \code{x_mat_new}. In addition, the function can provide a visualisation of the predictions in terms of the full posterior predictive distribution.
}
