% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.BayeSyneRgy.R
\name{summary.BayeSyneRgy}
\alias{summary.BayeSyneRgy}
\title{Summary function for BayeSyneRgy object}
\usage{
\method{summary}{BayeSyneRgy}(object, ...)
}
\arguments{
\item{object}{The fitted model to summarise, from \code{\link{BayeSyneRgy}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function summarizing posterior inference of a \code{BayeSyneRgy} object.
}
