% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesynergy3.R
\name{summary.bayesynergy3}
\alias{summary.bayesynergy3}
\title{Summary function for bayesynergy3 object}
\usage{
\method{summary}{bayesynergy3}(object, ...)
}
\arguments{
\item{object}{The fitted model to summarise, from \code{\link{bayesynergy3}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function summarizing posterior inference of a \code{bayesynergy3} object.
}
