% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.gibbs_BayeSyneRgy.R
\name{summary.gibbs_BayeSyneRgy}
\alias{summary.gibbs_BayeSyneRgy}
\title{Summary function for BayeSyneRgy object}
\usage{
\method{summary}{gibbs_BayeSyneRgy}(object, ...)
}
\arguments{
\item{object}{The fitted model to summarise, from \code{\link{BayeSyneRgy}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function summarizing posterior inference of a \code{BayeSyneRgy} object.
}
\examples{
library(BayeSyneRgy)
data("mathews_DLBCL")
y_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[1]]
x_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[2]]
fit <- BayeSyneRgy(y_mat,x_mat)
summary(fit)


}
