% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesynergy-package.R, R/bayesynergy.R
\docType{package}
\name{bayesynergy}
\alias{bayesynergy}
\title{bayesynergy}
\usage{
bayesynergy(
  y,
  x,
  type = 3,
  drug_names = NULL,
  experiment_ID = NULL,
  units = NULL,
  lower_asymptotes = F,
  heteroscedastic = T,
  lambda = 0.005,
  nu = 1.5,
  method = "sampling",
  control = list(),
  ...
)
}
\arguments{
\item{y}{vector or matrix of viability measures. Replicates can be given in long or wide format.}

\item{x}{two-column matrix of drug concentrations.}

\item{type}{integer; the type of model used. Must be one of the following: 1 (Splines), 2 (GP with squared exponential kernel), 3 (GP with Matérn kernel) or 4 (GP with rational quadratic kernel).}

\item{drug_names}{vector of size 2; names of the drugs utilized for the experiment.}

\item{experiment_ID}{character; identifier of experiment, typically name of cell Line.}

\item{units}{vector of size 2; concentration units for the drugs, e.g. c("\eqn{\mu}M","\eqn{\mu}M")}

\item{lower_asymptotes}{logical; if TRUE the model will estimate the lower asymptotes of monotherapy curves.}

\item{heteroscedastic}{logical; if TRUE, the model will assume heteroscedastic measurement error.}

\item{lambda}{numeric; the parameter controls the residual noise observed in the heteroscedastic model when f = 0.}

\item{nu}{numeric; the nu parameter for the Matérn kernel. Must be one of (0.5, 1.5, 2.5)}

\item{method}{The method of estimation. Must be one of {`sampling`,`vb`} corresponding to full sampling, or variational Bayes.}

\item{control}{list; passed on to the stan sampler, e.g. for setting adapt_delta.}

\item{...}{Arguments passed to \code{\link[rstan:sampling]{rstan::sampling}} or \code{\link[rstan:vb]{rstan::vb}} (e.g. iter, chains).}
}
\value{
An object of S3 class "\code{bayesynergy}", which is a list with the following entries
\tabular{ll}{
stanfit \tab An object of class \code{\link[rstan:stanmodel-class]{stanmodel}}, returned from the sampler. \cr
posterior_mean \tab A list containing the posterior means of model parameters. \cr
data \tab A list containing the original data used to fit the model. \cr
model \tab A list containing model specification for the model fit. \cr
returnCode \tab numeric; non-zero values indicate model was fit with errors or warnings \cr
LPML \tab numeric; The log pseudo-marginal likelihood of the fitted model. \cr
}
}
\description{
The function \code{bayesynergy} is the main function of the bayesynergy package. It will fit a Bayesian semi-parametric model for in-vitro drug combination experiments to estimate synergistic and antagonistic effects.
}
\examples{
library(bayesynergy)
data("mathews_DLBCL")
y_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[1]]
x_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[2]]
fit <- bayesynergy(y_mat,x_mat)

}
