% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayeSyneRgy-package.R, R/BayeSyneRgy.R
\docType{package}
\name{BayeSyneRgy}
\alias{BayeSyneRgy}
\title{BayeSyneRgy}
\usage{
BayeSyneRgy(
  y_mat,
  x_mat,
  drug_names = NULL,
  experiment_ID = NULL,
  log10_conc = FALSE,
  lower_asymptote = T,
  type = 2,
  Alg_param = list(),
  Hyper_param = list(),
  GP_param = list(),
  var_prior = 1,
  ...
)
}
\arguments{
\item{y_mat}{A matrix of responses, measured as share of viable cells after treatment. 
Each row corresponds to a unique concentration (x1,x2), while columns represents replicates.}

\item{x_mat}{A matrix of drug concentrations, each row corresponds to a unique concentration (x1,x2)
of the two drugs.}

\item{drug_names}{vector; a character vector containing the names of the drugs involved}

\item{experiment_ID}{character; name of cell line or other identifier of experiment}

\item{log10_conc}{logical; if true, concentrations must be given on the log10 scale, and monotherapies must have value -Inf for the drug with concentration zero.}

\item{lower_asymptote}{logical; if true, lower asymptotes are estimated for the monotherapies}

\item{type}{integer; the type of model used. Must be one of the following: 1 (Splines), 2 (GP with squared exponential kernel), 3 (GP with Matérn kernel) or 4 (GP with rational quadratic kernel).}

\item{Alg_param}{A list of parameters for the adaptive MCMC algorithm. See *Details* for more information.}

\item{Hyper_param}{A list of hyper parameters for the model. See *Details* for more information.}

\item{GP_param}{A list of parameters for the GP kernel function. See *Details* for more information.}

\item{var_prior}{Indicating which hyper-prior to put on variance terms. Either 1 (Inverse-Gamma) or 2 (Half-Cauchy)}

\item{...}{Additional parameters passed}
}
\value{
An object of S3 class “\code{BayeSyneRgy}”, which is a list with the following components:
\tabular{ll}{
OUTPUT \tab a list containing samples from the posterior distribution.  \cr 
data \tab the original data used to fit the model \cr
type \tab the type of model that was run. \cr
Alg_param \tab the parameters used for the MCMC sampler \cr
Hyper_param \tab the hyperparameters used for the prior distributions \cr
experiment_ID \tab the experiment identifier \cr
drug_names \tab the name of the drugs used in the experiment, taken from the column names of the concentration matrix. Used for plotting. \cr
Summary_Output \tab a list containing summary statistics calculated from the MCMC chains \cr
p_ij_mean \tab the response matrix \cr
p_0_mean \tab the calculated non-interaction matrix \cr
Delta_mean \tab the average posterior mean of the interaction \cr
}
}
\description{
The function \code{BayeSyneRgy} is the main function of the BayeSyneRgy package. It will fit a Bayesian semi-parametric model for in-vitro drug combination experiments to estimate synergistic and antagonistic effects. 

Open the package vignette for more details:

\code{vignette("BayeSyneRgy", package = "BayeSyneRgy")}
}
\details{
The argument \strong{Alg_param} is a list specifying any changes to the default values of the adaptive MCMC sampler. Any element passed will overwrite the defaults.
The list elements are:
\tabular{ll}{
 n_burn: \tab number of burn-in iterations for the MCMC sampler \cr
 thin: \tab amount of thinning \cr
 n_save: \tab number of final iterations returned to the user \cr
 g0: \tab adaptive burn-in \cr
 wg: \tab exponent of the adaptive function \cr 
 opt_rate: \tab optimal acceptance rate \cr
 eps: \tab parameter used for numerical stability \cr
}

The argument \strong{Hyper_param} specifies hyperparameters for the parametric part of the model. Any elements passed will overwrite the defaults.
All parameters are prefixed with either 'a_' and 'b_' for the Gamma or Inverse-Gamma option, or 'h_' for the Half-Cauchy prior. The table below gives the parameter names, list of valid priors, and a description of the parameter.
\tabular{lll}{
Slope_1: \tab G \tab Hyper-parameters for monotherapy slope of drug 1 \cr
Slope_2: \tab G \tab Hyper-parameters for monotherapy slope of drug 2 \cr
La_1: \tab G \tab Hyper-parameters for monotherapy lower asymptote of drug 1 \cr
La_2: \tab G \tab Hyper-parameters for monotherapy lower asymptote of drug 2 \cr
b1: \tab G \tab Hyper-parameters regulating the behaviour of the transformation g() \cr
b2: \tab G \tab Hyper-parameters regulating the behaviour of the transformation g() \cr
s2_Ec50_1: \tab IG,HC \tab Hyperparameters for the variance of the EC-50 for drug 1 \cr
s2_Ec50_2: \tab IG,HC \tab Hyperparameters for the variance of the EC-50 for drug 2 \cr
s2_gamma_0: \tab IG,HC \tab Hyperparameters for the variance of the gamma_0 parameter \cr
s2_gamma_1: \tab IG,HC \tab Hyperparameters for the variance of the gamma_1 parameter \cr
s2_gamma_2: \tab IG,HC \tab Hyperparameters for the variance of the gamma_2 parameter \cr
s2_eps: \tab IG,HC \tab Hyperparameters for the noise term \cr
}
For example, 'h_s2_eps' is a valid parameter name for a half-Cauchy prior on the noise term. This must be used in conjunction with \code{var_prior=2} for the half-cauchy option.

The argument \strong{GP_param} specifies hyperparamters for the Gaussian process covariance kernel. The elements are
\tabular{ll}{
ell: \tab The length scale used in all kernels. Depending on the length of the argument, either a fixed number, or a vector of length two giving the hyperparameters for an Inverse-Gamma prior distribution. \cr
nu: \tab The nu parameter for the Matérn kernel.  \cr
alpha: \tab The alpha parameter for the rational quadratic kernel. Depending on the length of the argument, either a fixed number, or a vector of length two giving the hyperparameters for an Inverse-Gamma prior distribution.  \cr
sigma2_f: \tab The kernel amplitude, i.e. \eqn{\sigma^2 k(x,x')}. Depending on the length of the argument and the specification in \code{sigma2_f_prior}, either a fixed number; a single hyperparameter for a Half-Cauchy prior; or hyperparameters for an Inverse-Gamma prior.  \cr
sigma2_f_prior: \tab An indicator for the type of prior on the kernel amplitude. Either 0 (fixed), 1 (Inverse-Gamma(a,b)) or 2 (Half-Cauchy(h))  \cr
}
}
\examples{
library(BayeSyneRgy)
data("mathews_DLBCL")
y_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[1]]
x_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[2]]
fit <- BayeSyneRgy(y_mat,x_mat)


}
\references{
Cremaschi A, Frigessi A, Taskén K, Zucknick M. (2019) A Bayesian approach for the study of synergistic interaction effects in in-vitro drug combination experiments. arXiv.org. https://arxiv.org/abs/1904.04901.
}
