% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesynergy.R
\name{bayesynergy}
\alias{bayesynergy}
\title{Bayesian semi-parametric modelling for in-vitro drug combination experiments}
\usage{
bayesynergy(
  y_mat,
  x_mat,
  drug_names = NULL,
  experiment_ID = NULL,
  log10_conc = FALSE,
  lower_asymptote = T,
  type = 2,
  ...
)
}
\description{
The function \code{bayesynergy} is the main function of the BayeSyneRgy package. It will fit a Bayesian semi-parametric model for in-vitro drug combination experiments to estimate synergistic and antagonistic effects.
}
