% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bayesynergy.R
\name{plot.bayesynergy}
\alias{plot.bayesynergy}
\title{Plot function for a bayesynergy object}
\usage{
\method{plot}{bayesynergy}(x, plot3D = T, save_plot = FALSE, path = NULL, plotdevice = "pdf", ...)
}
\arguments{
\item{x}{bayesynergy object, the result of \code{\link{bayesynergy}}.}

\item{plot3D}{logical; if TRUE interactive 3d surface plots of dose response function are displayed.}

\item{save_plot}{logical; if TRUE plots are saved locally.}

\item{path}{string; path for saving plots, if NULL defaults to work directory.}

\item{plotdevice}{string; device for saving plots locally, must be 'pdf' or 'png'}

\item{...}{further arguments passed on to device for plotting, useful for setting width and height of saved plots.}
}
\description{
A function for plotting synergy surfaces and summary statistics of a \code{bayesynergy} object.
}
\details{
This function extends \code{plot} to draw response and interaction surfaces of the fitted model. Both three-dimensional interactive plots, and two-dimensional contour plots can be displayed.
}
\examples{
library(BayeSyneRgy)
data("mathews_DLBCL")
y_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[1]]
x_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[2]]
fit <- bayesynergy(y_mat,x_mat)
plot(fit)

}
