% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesynergy.R
\name{summary.bayesynergy}
\alias{summary.bayesynergy}
\title{Summary function for bayesynergy object}
\usage{
\method{summary}{bayesynergy}(object, ...)
}
\arguments{
\item{object}{The fitted model to summarise, from \code{\link{bayesynergy}}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
A function summarizing posterior inference of a \code{bayesynergy} object.
}
\examples{
library(BayeSyneRgy)
data("mathews_DLBCL")
y_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[1]]
x_mat <- mathews_DLBCL$`ispinesib + ibrutinib`[[2]]
fit <- bayesynergy(y_mat,x_mat)
summary(fit)


}
